Adiciona express-validator e cria validação de user

+++ src/controllers/Users.js
@@ -1,3 +1,4 @@
+const { validationResult } = require('express-validator');
 const userRepository = require('../services/Users');
 
 async function get(req, res) {
@@ -20,6 +21,10 @@ async function getOne(req, res) {
 
 async function update(req, res) {
   try {
+    const validationErrors = validationResult(req);
+    if (!validationErrors.isEmpty()) {
+      return res.status(400).json(validationErrors.array());
+    }
     const result = await userRepository.update(req.params.id, req.body);
     return res.json(result);
   } catch (err) {
@@ -29,6 +34,10 @@ async function update(req, res) {
 
 async function insert(req, res) {
   try {
+    const validationErrors = validationResult(req);
+    if (!validationErrors.isEmpty()) {
+      return res.status(400).json(validationErrors.array());
+    }
     const user = req.body;
     const result = await userRepository.create(user);
     return res.json(result);
@@ -39,8 +48,12 @@ async function insert(req, res) {
 
 async function remove(req, res) {
   try {
+    const validationErrors = validationResult(req);
+    if (!validationErrors.isEmpty()) {
+      return res.status(404).json(validationErrors.array());
+    }
     const result = await userRepository.delete(req.params.id);
-    return (result.affected ? res.status(200).json(result) : res.status(404).json(result));
+    return res.json(result);
   } catch (err) {
     return res.status(400).json(err);
   }
